Rename config module to base
