+++ benchmarks/domains/__init__.py
@@ -1,6 +1,6 @@
 """Benchmark domains."""
 
-from benchmarks.definition.config import Benchmark
+from benchmarks.definition.base import Benchmark
 from benchmarks.domains.synthetic_2C1D_1C import synthetic_2C1D_1C_benchmark
 
 BENCHMARKS: list[Benchmark] = [
